% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_barplot.R
\name{generate_barplot}
\alias{generate_barplot}
\title{Create a bar plot}
\usage{
generate_barplot(dataset, x, x_name)
}
\arguments{
\item{dataset}{A data frame containing the categorical variable.}

\item{x}{The name of the column to be visualized (as a string).}

\item{x_name}{A user-friendly label for the x-axis.}
}
\value{
A ggplot2 object displaying the bar plot.
}
\description{
Create a bar plots for to visualize the distribution of a categorical variable in a data frame.
It ensures the input dataset is valid and that the selected column is a factor before plotting.
}
\examples{
# example code
barplot <- generate_barplot(dataset = mtcars, x = "gear", x_name = "Gear")

\dontrun{
barplot <- generate_barplot(dataset = mtcars, x = "gear", x_name = gear)

}
}
